+++ built_in/export_utils.c
@@ -6,35 +6,11 @@
 /*   By: junseyun <junseyun@student.42gyeongsan.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 20:44:49 by junseyun          #+#    #+#             */
-/*   Updated: 2024/11/23 19:13:36 by junseyun         ###   ########.fr       */
+/*   Updated: 2024/11/24 17:10:58 by junseyun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "env.h"
-
-void	add_exp_data(t_env_node *exp_list, char *data)
-{
-	t_env_node	*new_node;
-
-	new_node = create_node(data);
-	if (new_node != NULL)
-		add_node_back(&exp_list, new_node);
-	split_key_val(exp_list);
-}
-
-void	add_exp_env_data(t_env_node *exp_list, t_env_node *env_list, char *data)
-{
-	t_env_node	*new_node;
-
-	new_node = create_node(data);
-	if (new_node != NULL)
-	{
-		add_node_back(&exp_list, new_node);
-		add_node_back(&env_list, new_node);
-	}
-	new_node = 0;
-	split_key_val(exp_list);
-}
 
 int	cnt_equal(char *data)
 {
@@ -82,126 +58,6 @@ int	check_plus_operator(char *data)
 		i++;
 	}
 	return (0);
-}
-
-int	check_plus_operator_idx(char *data)
-{
-	int	i;
-
-	i = 0;
-	while (data[i] != '+')
-		i++;
-	return (i);
-}
-
-char	*create_env_data(char *data)
-{
-	int		i;
-	int		j;
-	int		len;
-	char	*new_str;
-
-	len = ft_strlen(data);
-	new_str = (char *)malloc(sizeof(char) * len);
-	i = 0;
-	j = 0;
-	while (data[i])
-	{
-		if (data[i] == '+')
-			i++;
-		else
-			new_str[j++] = data[i++];
-	}
-	new_str[j] = 0;
-	return (new_str);
-}
-
-void	add_new_exp_node(t_env_node **exp_list, char *data)
-{
-	char		*env_data;
-	t_env_node	*new_node;
-
-	env_data = create_env_data(data);
-	new_node = create_node(env_data);
-	if (new_node != NULL)
-		add_node_back(exp_list, new_node);
-}
-char	*split_key(char *data)
-{
-	int		i;
-	int		idx;
-	char	*key;
-
-	i = -1;
-	idx = check_plus_operator_idx(data);
-	key = (char *)malloc(sizeof(char) * (idx + 1));
-	while (++i < idx)
-		key[i] = data[i];
-	key[i] = 0;
-	return (key);
-}
-char	*split_value(char *data)
-{
-	int		i;
-	int		idx;
-	int		len;
-	char	*value;
-
-	i = 0;
-	idx = check_plus_operator_idx(data);
-	len = ft_strlen(data);
-	value = (char *)malloc(sizeof(char) * (len - idx - 1));
-	idx += 2;
-	while (data[idx])
-	{
-		value[i] = data[idx];
-		i++;
-		idx++;
-	}
-	value[i] = 0;
-	return (value);
-}
-void	add_export(t_env_node *exp_list, t_env_node *env_list, char *data)
-{
-	add_new_exp_node(&exp_list, data);
-	add_new_exp_node(&env_list, data);
-	set_split_exp_list(exp_list);
-}
-
-void	check_env_data(t_env_node *env_list, char *data)
-{
-	int			len;
-	int			flag;
-	char		*add_data;
-	t_env_node	*node;
-
-	node = env_list;
-	flag = 0;
-	add_data = create_env_data(data);
-	len = ft_strlen(add_data);
-	while (node != NULL)
-	{
-		if (ft_strncmp(node->env_data, add_data, len) == 0)
-		{
-			free(node->env_data);
-			node->env_data = add_data;
-			flag = 1;
-			break;
-		}
-		node = node -> next;
-	}
-	if (flag == 0)
-		add_new_exp_node(&env_list, data);
-}
-
-void	update_exp_node(t_env_node *node, char *key, char *value)
-{
-	char	*new_value;
-
-	new_value = ft_strjoin(node->value, value);
-	free_exp_key_value(node);
-	node->key = key;
-	node->value = new_value;
 }
 
 void	join_exp_data(t_env_node *exp_list, t_env_node *env_list, char *data)
@@ -252,7 +108,7 @@ void	cmd_export(t_token *node, t_env_node *exp_list, t_env_node *env_list)
 				else if (cnt_equal(temp->data) == 0)
 					add_exp_data(exp_list, temp->data);
 				else if (check_plus_operator(temp->data))
-					join_exp_data(exp_list, env_list ,temp->data);
+					join_exp_data(exp_list, env_list, temp->data);
 				else if (check_validation(temp->data) == 1)
 					add_exp_env_data(exp_list, env_list, temp->data);
 				temp = temp -> next;
